+++ src/palette_util.c
@@ -103,7 +103,7 @@ static u8 RouletteFlash_FadePalette(struct RouletteFlashPalette *pal)
 
 static u8 RouletteFlash_FlashPalette(struct RouletteFlashPalette *pal)
 {
-    u8 i = 0;
+    u32 i = 0;
     switch (pal->state)
     {
     case 1:
@@ -124,7 +124,7 @@ static u8 RouletteFlash_FlashPalette(struct RouletteFlashPalette *pal)
 
 void RouletteFlash_Run(struct RouletteFlashUtil *flash)
 {
-    u8 i = 0;
+    u32 i = 0;
 
     if (flash->enabled)
     {
@@ -148,7 +148,7 @@ void RouletteFlash_Run(struct RouletteFlashUtil *flash)
 
 void RouletteFlash_Enable(struct RouletteFlashUtil *flash, u16 flags)
 {
-    u8 i = 0;
+    u32 i = 0;
 
     flash->enabled++;
     for (i = 0; i < ARRAY_COUNT(flash->palettes); i++)
@@ -206,7 +206,7 @@ void RouletteFlash_Stop(struct RouletteFlashUtil *flash, u16 flags)
 
 void InitPulseBlend(struct PulseBlend *pulseBlend)
 {
-    u8 i = 0;
+    u32 i = 0;
     pulseBlend->usedPulseBlendPalettes = 0;
     memset(&pulseBlend->pulseBlendPalettes, 0, sizeof(pulseBlend->pulseBlendPalettes));
     for (; i < 16; i++)
@@ -215,7 +215,7 @@ void InitPulseBlend(struct PulseBlend *pulseBlend)
 
 int InitPulseBlendPaletteSettings(struct PulseBlend *pulseBlend, const struct PulseBlendSettings *settings)
 {
-    u8 i = 0;
+    u32 i = 0;
     struct PulseBlendPalette *pulseBlendPalette = NULL;
 
     if (!pulseBlend->pulseBlendPalettes[0].inUse)
@@ -269,7 +269,7 @@ static void ClearPulseBlendPalettesSettings(struct PulseBlendPalette *pulseBlend
 
 void UnloadUsedPulseBlendPalettes(struct PulseBlend *pulseBlend, u16 pulseBlendPaletteSelector, u8 multiSelection)
 {
-    u16 i = 0;
+    u32 i = 0;
 
     if (!multiSelection)
     {
@@ -289,7 +289,7 @@ void UnloadUsedPulseBlendPalettes(struct PulseBlend *pulseBlend, u16 pulseBlendP
 
 void MarkUsedPulseBlendPalettes(struct PulseBlend *pulseBlend, u16 pulseBlendPaletteSelector, u8 multiSelection)
 {
-    u8 i = 0;
+    u32 i = 0;
 
     if (!multiSelection)
     {
@@ -318,7 +318,7 @@ void UnmarkUsedPulseBlendPalettes(struct PulseBlend *pulseBlend, u16 pulseBlendP
 {
     u32 i;
     struct PulseBlendPalette *pulseBlendPalette;
-    u8 j = 0;
+    u32 j = 0;
 
     if (!multiSelection)
     {
@@ -362,7 +362,7 @@ void UnmarkUsedPulseBlendPalettes(struct PulseBlend *pulseBlend, u16 pulseBlendP
 void UpdatePulseBlend(struct PulseBlend *pulseBlend)
 {
     struct PulseBlendPalette *pulseBlendPalette;
-    u8 i = 0;
+    u32 i = 0;
 
     if (pulseBlend->usedPulseBlendPalettes)
     {
