+++ src/rayquaza_scene.c
@@ -1642,7 +1642,7 @@ static void Task_DuoFightAnim(u8 taskId)
 
 static void Task_DuoFight_AnimateClouds(u8 taskId)
 {
-    s16 i;
+    s32 i;
     u16 *data = (u16*)gTasks[taskId].data;
 
     for (i = 24; i < 92; i++)
